# repo/api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: KuberCloud ANI Core API
  version: 1.0.0
  description: |
    KuberCloud ANI Core — 基础设施平台 REST API 契约（唯一真实来源）。

    ## 文件范围与边界
    本文件（api/openapi/v1.yaml）仅包含 ANI Core 基础设施资源：
    instances（VM/Container/GPU/Sandbox/BM/K8s集群）、volumes/filesystems/objects/
    vector-stores、networks、auth/secrets/encryption、gpu-inventory/dpu-inventory、
    registry/metering/observability/notifications/audit 等。
    ANI Services 资源（模型仓库/推理/知识库/PaaS）在 api/openapi/services/v1.yaml。

    ## URL 前缀规范
    servers[0].url = "https://{host}/api/v1"（前缀在 spec 中定义，Gateway 不做路径变换）。
    版本升级时新建 api/openapi/v2.yaml，本文件路径保持稳定。

    ## SDK 生成（make gen-core-sdk）
    Go: github.com/kubercloud/ani-go | Python: pip install kubercloud-ani
    TypeScript: @kubercloud/ani-client | Java: com.kubercloud:ani-java

    ## 认证
    Bearer JWT（1小时有效）或 X-API-Key Header。
    tenant_id 从 JWT claims 提取，请求体中的 tenant_id 字段将被忽略。

    ## 错误格式
    `{ code: string, message: string, request_id: string, details?: object }`
    标准错误码：UNAUTHORIZED / FORBIDDEN / NOT_FOUND / CONFLICT / BAD_REQUEST /
    RATE_LIMIT_EXCEEDED / NOT_IMPLEMENTED / INTERNAL_ERROR

    ## 分页
    cursor 分页：`?limit=20&cursor=<token>` → `{ items: [], next_cursor: string | null }`

    ## 异步操作
    202 响应体返回 AsyncTask，Location Header 附任务 URL。
    轮询 GET /api/v1/tasks/{task_id} 或配置 Webhook。
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: ani.internal
        description: ANI Gateway 地址（支持IP地址）

...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT AccessToken（有效期1小时，通过 /auth/token 获取）
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 长期 API Key（在 Console → 设置 → API Key 中创建）

# ── 公共 Schema ───────────────────────────────────────────────────────────────
  schemas:
    K8sClusterCreateRequest:
      type: object
      properties:
        idempotency_key: { type: string }
        name:            { type: string }
        version:         { type: string }
      required: [idempotency_key, name]
    K8sClusterUpgradeRequest:
      type: object
      properties:
        idempotency_key: { type: string }
        version:         { type: string }
      required: [idempotency_key, version]
    K8sCluster:
      type: object
      properties:
        id:         { type: string }
        tenant_id:  { type: string }
        name:       { type: string }
        version:    { type: string }
        state:      { type: string, enum: [provisioning, running, deleting] }
        reason:     { type: string }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    K8sClusterListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/K8sCluster" } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    K8sClusterNodePoolGPU:
      type: object
      properties:
        vendor:        { type: string }
        model:         { type: string }
        count:         { type: integer }
        resource_name: { type: string }
    K8sClusterNodePoolCreateRequest:
      type: object
      properties:
        idempotency_key: { type: string }
        name:            { type: string }
        node_count:      { type: integer, minimum: 1 }
        instance_type:   { type: string }
        gpu:
          $ref: '#/components/schemas/K8sClusterNodePoolGPU'
      required: [idempotency_key, name, node_count, instance_type]
    K8sClusterNodePoolUpdateRequest:
      type: object
      properties:
        idempotency_key: { type: string }
        node_count:      { type: integer, minimum: 0 }
        instance_type:   { type: string }
        gpu:
          $ref: '#/components/schemas/K8sClusterNodePoolGPU'
      required: [idempotency_key, node_count, instance_type]
    K8sClusterNodePool:
      type: object
      properties:
        id:            { type: string }
        tenant_id:     { type: string }
        cluster_id:    { type: string }
        name:          { type: string }
        node_count:    { type: integer }
        instance_type: { type: string }
        gpu:
          $ref: '#/components/schemas/K8sClusterNodePoolGPU'
        state:         { type: string, enum: [running, deleting] }
        reason:        { type: string }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at:    { type: string, format: date-time }
        updated_at:    { type: string, format: date-time }
    K8sClusterNodePoolListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/K8sClusterNodePool" } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    K8sClusterKubeconfig:
      type: object
      properties:
        cluster_id: { type: string }
        tenant_id:  { type: string }
        server:     { type: string }
        namespace:  { type: string }
        ca_data:    { type: string }
        token:      { type: string }
        kubeconfig: { type: string }
        expires_at: { type: string, format: date-time }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
    K8sClusterProxyRequest:
      type: object
      properties:
        idempotency_key: { type: string }
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        path:
          type: string
          description: 原生 Kubernetes API path，仅允许 /api/、/apis/、/healthz、/livez、/readyz、/version 范围。
        query:
          type: object
          additionalProperties: { type: string }
        body:
          type: object
          additionalProperties: true
      required: [idempotency_key, method, path]
    K8sClusterProxyResponse:
      type: object
      properties:
        cluster_id: { type: string }
        tenant_id:  { type: string }
        method:     { type: string }
        path:       { type: string }
        query:
          type: object
          additionalProperties: { type: string }
        status_code: { type: integer }
        headers:
          type: object
          additionalProperties: { type: string }
        body:
          type: object
          additionalProperties: true
        proxied_at: { type: string, format: date-time }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:        { type: string, example: "MODEL_NOT_FOUND" }
        message:     { type: string, example: "模型 qwen2.5-72b 不存在" }
        request_id:  { type: string, example: "req_01j8xabc" }
        details:     { type: object, additionalProperties: true }

    # Cursor 分页（所有列表接口统一使用此格式）
...
    CoreDevProfileInfo:
      type: object
      description: "Core dev/local profile 标记；用于区分本地联调成功与真实 provider 执行成功。"
      required: [mode, provider, real_provider]
      properties:
        mode:          { type: string, enum: [local, real] }
        provider:      { type: string }
        real_provider: { type: boolean }
        reason:        { type: string, nullable: true }
...
  responses:
...
    NotFound:
      description: 资源不存在（code=NOT_FOUND）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: 请求参数不合法（code=BAD_REQUEST）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 资源已存在或冲突（code=CONFLICT）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ── API 路径定义 ──────────────────────────────────────────────────────────────
# 说明：/demo/* 路径已迁移至 api/openapi/dev.yaml（不参与 SDK 生成）
paths:
...
  /k8s-clusters:
    get:
      operationId: listK8sClusters
      summary: 列出 K8s 集群
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:read"
      parameters:
        - { name: limit,  in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sClusterListResponse"
    post:
      operationId: createK8sCluster
      summary: 创建 K8s 集群
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:create"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/K8sClusterCreateRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sCluster"
  /k8s-clusters/{cluster_id}:
    get:
      operationId: getK8sCluster
      summary: 查询 K8s 集群
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:read"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sCluster"
    delete:
      operationId: deleteK8sCluster
      summary: 删除 K8s 集群
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:delete"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sCluster"
  /k8s-clusters/{cluster_id}/kubeconfig:
    get:
      operationId: getK8sClusterKubeconfig
      summary: 获取 K8s 集群 kubeconfig
      description: Sprint 5 local dev profile 返回模拟 vCluster kubeconfig；真实 provider 接入后必须保持响应语义兼容。
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:read"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sClusterKubeconfig"
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
  /k8s-clusters/{cluster_id}/upgrade:
    post:
      operationId: upgradeK8sCluster
      summary: 升级 K8s 集群版本
      description: Sprint 5 code boundary 会更新 Core cluster version，并在 vCluster provider mode 下委托 provider 表达 Helm upgrade 意图；真实 live 升级验证由 REAL-K8S-LAB-A / vCluster live gate 单独证明。
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:manage"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/K8sClusterUpgradeRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sCluster"
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
  /k8s-clusters/{cluster_id}/node-pools:
    get:
      operationId: listK8sClusterNodePools
      summary: 列出 K8s 集群节点池
      description: Sprint 5 local dev profile 返回租户集群下的节点池列表；真实节点池扩缩容验证由后续真实 provider/live gate 单独证明。
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:read"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sClusterNodePoolListResponse"
        "404":
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createK8sClusterNodePool
      summary: 创建 K8s 集群节点池
      description: Sprint 5 local dev profile 记录节点池规格、节点数和 GPU intent；真实节点池 provider apply/live 验证待后续切片。
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:manage"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/K8sClusterNodePoolCreateRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sClusterNodePool"
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
  /k8s-clusters/{cluster_id}/node-pools/{node_pool_id}:
    get:
      operationId: getK8sClusterNodePool
      summary: 获取 K8s 集群节点池
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:read"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
        - { name: node_pool_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sClusterNodePool"
        "404":
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateK8sClusterNodePool
      summary: 更新 K8s 集群节点池
      description: 用于调整节点数、实例规格和 GPU intent；`idempotency_key` 必须在客户端重试时保持不变。
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:manage"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
        - { name: node_pool_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/K8sClusterNodePoolUpdateRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sClusterNodePool"
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteK8sClusterNodePool
      summary: 删除 K8s 集群节点池
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:manage"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
        - { name: node_pool_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sClusterNodePool"
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
  /k8s-clusters/{cluster_id}/proxy:
    post:
      operationId: proxyK8sClusterAPI
      summary: 代理访问 K8s 集群原生 API
      description: Sprint 5 local dev profile 返回模拟 proxy 响应；真实 provider 接入后由 Core adapter 转发到租户 vCluster API Server，Gateway handler 不直接调用 K8s SDK。
      tags: [K8sClusters]
      x-ani-rbac-scope: "scope:k8s-clusters:proxy"
      parameters:
        - { name: cluster_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/K8sClusterProxyRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sClusterProxyResponse"
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
...